requirements.txt
astroid==1.6.6
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
configparser==3.7.4
decorator==4.4.0
Django==1.11.21
django-cors-headers==3.0.2
django-webpack-loader==0.6.0
djangorestframework==3.9.4
enum34==1.1.6
futures==3.2.0
html-validator==1.0.1
ipdb==0.12
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
ply==3.11
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.4.2
pylint==1.9.4
pytz==2019.1
scandir==1.10.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
tinycss==0.4
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.0
wrapt==1.11.1
